Bump Pinned Dependencies (#111)
* Bump requirements

* Remove OpenEye Toolkits from requirements. Still the suggested Charge
  method, but not available to me

* Bump C++ dependencies

* Pybind11 + Eigen 5

* Bump Mypy

* Fixes up some minor type issues

* Go back to OpenMM 8.3.1

* Fix tests

* Update hashes

* JAX update changes version

* Fix test

+++ tmd/fe/rbfe.py
@@ -276,7 +276,7 @@ def setup_initial_state(
             st, host, ligand_conf, lamb, temperature, init_seed, verify_constraints=verify_constraints
         )
         box0 = host.box
-        protein_idxs = np.arange(0, len(host.masses) - host.num_water_atoms)
+        protein_idxs = np.arange(0, len(host.masses) - host.num_water_atoms, dtype=np.int32)
     else:
         x0, box0, hmr_masses, potentials, baro = setup_in_vacuum(
             st, ligand_conf, lamb, verify_constraints=verify_constraints
